[refactor] #50 auth pakage 생성
