+++ app/src/main/java/org/android/bbangzip/data/auth/service/KakaoAuthService.kt
@@ -1,6 +1,6 @@
-package org.android.bbangzip.data.datasource.remote.service
+package org.android.bbangzip.data.auth.service
 
-import android.content.ContentValues.TAG
+import android.content.ContentValues
 import android.content.Context
 import com.kakao.sdk.auth.model.OAuthToken
 import com.kakao.sdk.user.UserApiClient
@@ -23,21 +23,21 @@ class KakaoAuthService
             val callback: (OAuthToken?, Throwable?) -> Unit = { oAuthToken, throwable ->
                 if (throwable != null) {
                     loginError(throwable = throwable, context = context)
-                    Timber.d("[카카오 로그인] -> 사용자 정보 요청 실패 $throwable")
+                    Timber.Forest.d("[카카오 로그인] -> 사용자 정보 요청 실패 $throwable")
                 } else if (oAuthToken != null) {
                     loginSuccess(
                         oAuthToken = oAuthToken,
                         loginListener = loginListener,
                     )
-                    Timber.d("[카카오 로그인] -> 사용자 정보 요청 성공 $oAuthToken")
+                    Timber.Forest.d("[카카오 로그인] -> 사용자 정보 요청 성공 $oAuthToken")
                 }
             }
 
             if (isKakaoTalkLoginAvailable(context)) {
-                Timber.d("[카카오 로그인] -> loginWithKakaoTalk $isKakaoTalkLoginAvailable")
+                Timber.Forest.d("[카카오 로그인] -> loginWithKakaoTalk $isKakaoTalkLoginAvailable")
                 client.loginWithKakaoTalk(context = context, callback = callback)
             } else {
-                Timber.d("[카카오 로그인] -> loginWithKakaoAccount $isKakaoTalkLoginAvailable")
+                Timber.Forest.d("[카카오 로그인] -> loginWithKakaoAccount $isKakaoTalkLoginAvailable")
                 client.loginWithKakaoAccount(context = context, callback = callback)
             }
         }
@@ -47,7 +47,7 @@ class KakaoAuthService
             context: Context,
         ) {
             val kakaoType = if (isKakaoTalkLoginAvailable(context)) KAKAO_TALK else KAKAO_ACCOUNT
-            Timber.d("[카카오 로그인] -> {$kakaoType}으로 로그인 실패 ${throwable.message}")
+            Timber.Forest.d("[카카오 로그인] -> {$kakaoType}으로 로그인 실패 ${throwable.message}")
         }
 
         private fun loginSuccess(
@@ -57,29 +57,29 @@ class KakaoAuthService
             client.me { _, error ->
                 loginListener(oAuthToken.accessToken)
                 if (error != null) {
-                    Timber.d("[카카오 로그인] -> 사용자 정보 요청 실패 $error")
+                    Timber.Forest.d("[카카오 로그인] -> 사용자 정보 요청 실패 $error")
                 }
             }
         }
 
         fun logoutKakao(logoutListener: () -> Unit) {
-            UserApiClient.instance.logout { error ->
+            UserApiClient.Companion.instance.logout { error ->
                 if (error != null) {
-                    Timber.tag(TAG).e(error.message, "[카카오 로그인] -> 카카오 로그아웃 실패.")
+                    Timber.Forest.tag(ContentValues.TAG).e(error.message, "[카카오 로그인] -> 카카오 로그아웃 실패.")
                 } else {
                     logoutListener()
-                    Timber.tag(TAG).i("[카카오 로그인] -> 카카오 로그아웃 성공.")
+                    Timber.Forest.tag(ContentValues.TAG).i("[카카오 로그인] -> 카카오 로그아웃 성공.")
                 }
             }
         }
 
         fun withdrawKakao(withdrawListener: () -> Unit) {
-            UserApiClient.instance.unlink { error ->
+            UserApiClient.Companion.instance.unlink { error ->
                 if (error != null) {
-                    Timber.tag(TAG).e(error, "[카카오 로그인] -> 카카오 회원 탈퇴 실패 : ${error.message}")
+                    Timber.Forest.tag(ContentValues.TAG).e(error, "[카카오 로그인] -> 카카오 회원 탈퇴 실패 : ${error.message}")
                 } else {
                     withdrawListener()
-                    Timber.tag(TAG).i("[카카오 로그인] -> 카카오 회원 탈퇴 성공. ")
+                    Timber.Forest.tag(ContentValues.TAG).i("[카카오 로그인] -> 카카오 회원 탈퇴 성공. ")
                 }
             }
         }
